Unreviewed. Speculative build fix for watchOS build.
* runtime/ArrayPrototype.cpp:
(JSC::shift):



Canonical link: https://commits.webkit.org/224186@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261005 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/runtime/ArrayPrototype.cpp
@@ -334,7 +334,7 @@ void shift(JSGlobalObject* globalObject, JSObject* thisObj, uint64_t header, uin
 
     if (isJSArray(thisObj)) {
         JSArray* array = asArray(thisObj);
-        uint32_t header32 = header;
+        uint32_t header32 = static_cast<uint32_t>(header);
         ASSERT(header32 == header);
         if (array->length() == length && array->shiftCount<shiftCountMode>(globalObject, header32, count))
             return;
